+++ fake-cucumber/javascript/src/types.ts
@@ -1,0 +1,3 @@
+import { messages } from 'cucumber-messages'
+
+export type MessageNotifier = (message: messages.IEnvelope) => void