+++ html-formatter/javascript/test/dummy_test.ts
@@ -1,3 +1,4 @@
 describe('cucumber-html-formatter', () => {
-  it('has no unit tests, but it has acceptance tests (see Makefile)', () => null)
+  it('has no unit tests, but it has acceptance tests (see Makefile)', () =>
+    null)
 })