+++ json-formatter/ruby/features/step_definitions/steps.rb
@@ -1,17 +1,18 @@
-Given("a passed step") do
+Given("a passed {word}") do |word|
 end
 
-Given("a passed step with") do |arg|
+Given("a passed {word} with") do |word, arg|
 end
 
-When("a failed step") do
+When("a failed {word}") do |word|
   raise "Some error"
 end
 
-When("a failed step with") do |arg|
+When("a failed {word} with") do |word, arg|
   raise "Some error"
 end
 
-Then("a skipped step") do
+Then("a skipped {word}") do |word|
+  raise "Should never be called"
 end