+++ cucumber-expressions/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* [JavaScript] export `Argument` and `Expression` types
+
 ### Changed
 
 ### Deprecated